requirements.txt
streamlit>=1.22.0
markdown>=3.4.3
fpdf2>=2.7.4
